motos_backend/requirements.txt
Django==6.0.3
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.0
Pillow>=10.0.0
psycopg[binary]>=3.1.0
